Entrega 3

+++ src/main/java/com/alkemy/disney/disney/service/impl/PeliculaServiceImpl.java
@@ -2,9 +2,11 @@ package com.alkemy.disney.disney.service.impl;
 
 import com.alkemy.disney.disney.dto.PeliculaBasicDTO;
 import com.alkemy.disney.disney.dto.PeliculaDTO;
+import com.alkemy.disney.disney.dto.PeliculaFiltersDTO;
 import com.alkemy.disney.disney.entity.Pelicula;
 import com.alkemy.disney.disney.mapper.PeliculaMapper;
 import com.alkemy.disney.disney.repository.PeliculaRepository;
+import com.alkemy.disney.disney.repository.specifications.PeliculaSpecifications;
 import com.alkemy.disney.disney.service.PeliculaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,6 +23,9 @@ public class PeliculaServiceImpl implements PeliculaService {
 
     @Autowired
     private PeliculaRepository peliculaRepository;
+
+    @Autowired
+    private PeliculaSpecifications peliculaSpecifications;
 
     @Override
     @Transactional
@@ -49,8 +54,13 @@ public class PeliculaServiceImpl implements PeliculaService {
 
     @Override
     @Transactional
-    public void delete(Long id) {
-        peliculaRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if (peliculaRepository.existsById(id)){
+            peliculaRepository.deleteById(id);
+            return true;
+        }else{
+            return false;
+        }
     }
 
     @Override
@@ -60,7 +70,7 @@ public class PeliculaServiceImpl implements PeliculaService {
         peliculaMapper.peliculaRefreshValues(respuesta.get(),peliculaDTO);
         Pelicula peliculaSaved = peliculaRepository.save(respuesta.get());
         PeliculaDTO result = peliculaMapper.pelicula2DTO(peliculaSaved);
-        return null;
+        return result;
     }
 
     @Override
@@ -69,5 +79,13 @@ public class PeliculaServiceImpl implements PeliculaService {
         Optional<Pelicula> respuesta = peliculaRepository.findById(id);
         PeliculaDTO result = peliculaMapper.pelicula2DTO(respuesta.get());
         return result;
+    }
+
+    @Override
+    public List<PeliculaDTO> getByFilters(String name, String genre, String order) {
+        PeliculaFiltersDTO filtersDTO = new PeliculaFiltersDTO(name, genre, order);
+        List<Pelicula> peliculaList = peliculaRepository.findAll(peliculaSpecifications.getByFilters(filtersDTO));
+        List<PeliculaDTO> peliculaDTOS = peliculaMapper.peliculaList2DTOList(peliculaList);
+        return peliculaDTOS;
     }
 }
